+++ src/app/Scenes/SubmitArtwork/UploadPhotos/utils/addAssetToConsignment.ts
