+++ src/app/Scenes/SubmitArtwork/UploadPhotos/utils/consignmentSetupToSubmission.ts
@@ -1,7 +1,7 @@
 import { CreateSubmissionMutationInput } from "__generated__/createConsignmentSubmissionMutation.graphql"
 import { UpdateSubmissionMutationInput } from "__generated__/updateConsignmentSubmissionMutation.graphql"
 import { isNil, omitBy } from "lodash"
-import { ConsignmentSetup } from "../index"
+import { ConsignmentSetup } from "./index"
 
 export const consignmentSetupToMutationInput = (submission: ConsignmentSetup) =>
   omitBy(
