+++ src/app/Scenes/SubmitArtwork/UploadPhotos/utils/gemini/createGeminiAssetWithS3Credentials.ts
