+++ src/app/Scenes/SubmitArtwork/UploadPhotos/utils/gemini/getGeminiCredentialsForEnvironment.ts
