+++ src/app/Scenes/SubmitArtwork/UploadPhotos/utils/index.tests.tsx
