+++ src/app/Scenes/SubmitArtwork/UploadPhotos/utils/uploadPhotoToGemini.ts
